requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==23.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cog==0.13.7
fastapi==0.115.8
h11==0.14.0
httptools==0.6.4
idna==3.10
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.45.3
structlog==24.4.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
cog
Pillow>=8.0.0

